+++ src/components/App.js
@@ -14,7 +14,7 @@ function App() {
   return (
     <div className="container mx-auto grid grid-cols-6 gap-4 mt-4">
       <NavBar />
-      <div className="ontoggle">
+      <div className="ontoggle ontoggle-close">
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/login" element={<LoginForm />} />
